test: Extract `decorator_python_version`
