+++ test/integ_tests/test_reservation_arn.py
@@ -15,12 +15,12 @@ import sys
 
 import pytest
 from botocore.exceptions import ClientError
+from job_testing_utils import decorator_python_version
 
 from braket.aws import AwsDevice, DirectReservation
 from braket.circuits import Circuit
 from braket.devices import Devices
 from braket.jobs import get_job_device_arn, hybrid_job
-from braket.test.integ_tests.test_create_quantum_job import decorator_python_version
 
 
 @pytest.fixture
